Moved linked list functions from areas to linked_list
